Add documentation around content data structure
Initial documentation is added around how to explore the site's data
and content using GraphQL. The example markdown page has been expanded
to include more examples.

As we move forwards with this we alwso need to consider more complex
requirements including custom components and image embedding. The ways
in which we decide to meet these requirements will influence how we
structure the markdown files as time goes on.