fix: use astro styles for social icons
Emotion isn’t working properly, so let’s keep styles native to Astro.
Styles are scoped by default, so Emotion is not adding much.

Using Source is not a priority, or much of an aim. As long as
this implementation is as accessible as Source’s, we are fine.

We can remove unused icons component, and only use the icons as SVG.